컴포넌트 화
