+++ src/screens/LocationsScreen.tsx
@@ -1,105 +1,17 @@
 import React from 'react';
-import {
-  SafeAreaView,
-  StyleSheet,
-  Text,
-  TextInput,
-  TouchableOpacity,
-  View,
-} from 'react-native';
-import Animated, {
-  useAnimatedStyle,
-  useSharedValue,
-  withTiming,
-} from 'react-native-reanimated';
-import MortyImage from '../assets/images/morty.svg';
-import RickImage from '../assets/images/rick.svg';
-
-const DEFAULT_LEFT_VALUE = 0;
-const DEFAULT_DEGREE = '0';
+import {SafeAreaView, StyleSheet, Text, View} from 'react-native';
+import MoveRick from '../components/animation/MoveRick';
+import RotateMorty from '../components/animation/RotateMorty';
 
 const LocationsScreen = () => {
-  // move Rick
-  const leftValue = useSharedValue<number>(DEFAULT_LEFT_VALUE);
-  const [rickMoveValue, setRickMoveValue] = React.useState('0');
-  const handleMoveValue = (num: string) => {
-    setRickMoveValue(num);
-  };
-  const moveAnimatedStyle = useAnimatedStyle(() => {
-    return {
-      left: leftValue.value,
-    };
-  });
-  const moveRick = (num: string) => {
-    leftValue.value = withTiming(isNaN(Number(num)) ? 0 : parseInt(num));
-    setRickMoveValue('0');
-  };
-
-  // rotate Morty
-  const degree = useSharedValue<string>(DEFAULT_DEGREE);
-  const [mortyRotateValue, setMortyRotateValue] = React.useState('0');
-  const handleRotateValue = (degree: string) => {
-    setMortyRotateValue(degree);
-  };
-  const rotateAnimatedStyle = useAnimatedStyle(() => {
-    return {
-      transform: [{rotate: `${degree.value}deg`}],
-    };
-  });
-
-  const rotateMorty = (deg: string) => {
-    degree.value = withTiming(isNaN(Number(deg)) ? '0' : deg, {duration: 1500});
-    setMortyRotateValue('0');
-  };
-
   return (
     <SafeAreaView style={styles.safeArea}>
       <View style={styles.wrapper}>
         <View>
           <Text style={styles.header}>Learn Animation</Text>
         </View>
-        <View style={styles.itemWrapper}>
-          <View style={styles.controlBoxWrapper}>
-            <TextInput
-              style={styles.textInput}
-              onChangeText={handleMoveValue}
-              value={rickMoveValue}
-              placeholder="0"
-              placeholderTextColor={'#000000'}
-            />
-            <TouchableOpacity
-              style={styles.moveButton}
-              onPress={() => moveRick(rickMoveValue)}>
-              <Text>Move</Text>
-            </TouchableOpacity>
-          </View>
-          <View style={styles.borderBox}>
-            <Animated.View style={[styles.imageBox, moveAnimatedStyle]}>
-              <RickImage width={50} height={50} />
-            </Animated.View>
-          </View>
-        </View>
-        <View style={styles.itemWrapper}>
-          <View style={styles.controlBoxWrapper}>
-            <TextInput
-              style={styles.textInput}
-              onChangeText={handleRotateValue}
-              value={mortyRotateValue}
-              placeholder="0"
-              placeholderTextColor={'#000000'}
-            />
-            <TouchableOpacity
-              style={styles.rotateButton}
-              onPress={() => rotateMorty(mortyRotateValue)}>
-              <Text>Rotate</Text>
-            </TouchableOpacity>
-          </View>
-          <View>
-            <Animated.View style={[styles.imageBox, rotateAnimatedStyle]}>
-              <MortyImage width={50} height={50} />
-            </Animated.View>
-          </View>
-        </View>
+        <MoveRick />
+        <RotateMorty />
       </View>
     </SafeAreaView>
   );
@@ -122,51 +34,5 @@ const styles = StyleSheet.create({
     color: '#FFFFFF',
     fontSize: 24,
     fontWeight: '700',
-  },
-  itemWrapper: {gap: 10},
-  controlBoxWrapper: {
-    flexDirection: 'row',
-    gap: 10,
-  },
-  textInput: {
-    backgroundColor: '#FFFFFF',
-    width: 63,
-    borderRadius: 4,
-    fontWeight: '700',
-    fontSize: 16,
-    textAlign: 'center',
-  },
-  moveButton: {
-    backgroundColor: '#8CD790',
-    borderRadius: 4,
-    width: 64,
-    height: 30,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  rotateButton: {
-    backgroundColor: '#85BFE9',
-    borderRadius: 4,
-    width: 64,
-    height: 30,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  buttonText: {
-    color: '#FFFFFF',
-    fontSize: 12,
-  },
-  borderBox: {
-    height: 80,
-    borderRadius: 14,
-    borderColor: '#FFFFFF',
-    borderWidth: 2,
-    justifyContent: 'center',
-  },
-  imageBox: {
-    width: 64,
-    height: 64,
-    justifyContent: 'center',
-    alignItems: 'center',
   },
 });
